finished refactoring everything so that missing columns btwn college and nba players are handled properly
