+++ recovery_files/early_scripts/compare.py
