+++ recovery_files/early_scripts/pull_college_player_headshot.py
