+++ recovery_files/early_scripts/scrape_nba_height_image_link
