+++ templates/index.html
@@ -164,7 +164,7 @@
             <br>
 
             <span style="color: white; font-weight: bold; text-align: center; display: block; font-family: 'Oswald', sans-serif; margin-top: 15px;">
-                Please allow up to 15 seconds for your comparison to load...
+                Please allow up to 30 seconds for your comparison to load...
               </span>
 
         </div>
